+++ website/sidebars.js
@@ -125,6 +125,11 @@ module.exports = {
         },
         {
           "type": "doc",
+          "id": "cookbook/router",
+          "label": "Router"
+        },
+        {
+          "type": "doc",
           "id": "cookbook/control-refs",
           "label": "Control Refs"
         },
@@ -970,6 +975,10 @@ module.exports = {
                     },
                     {
                       "type": "doc",
+                      "id": "controls/map/overlayimagelayer"
+                    },
+                    {
+                      "type": "doc",
                       "id": "controls/map/circlelayer"
                     },
                     {
@@ -1002,6 +1011,25 @@ module.exports = {
                     {
                       "type": "doc",
                       "id": "controls/map/polylinemarker"
+                    }
+                  ]
+                },
+                {
+                  "type": "category",
+                  "label": "Overlays",
+                  "collapsed": true,
+                  "items": [
+                    {
+                      "type": "doc",
+                      "id": "controls/map/baseoverlayimage"
+                    },
+                    {
+                      "type": "doc",
+                      "id": "controls/map/overlayimage"
+                    },
+                    {
+                      "type": "doc",
+                      "id": "controls/map/rotatedoverlayimage"
                     }
                   ]
                 },
@@ -1190,6 +1218,11 @@ module.exports = {
               "type": "doc",
               "id": "controls/rotatedbox",
               "label": "RotatedBox"
+            },
+            {
+              "type": "doc",
+              "id": "controls/router",
+              "label": "Router"
             },
             {
               "type": "doc",
@@ -2162,6 +2195,10 @@ module.exports = {
                   "id": "types/localeconfiguration"
                 },
                 {
+                  "type": "doc",
+                  "id": "types/locationinfo"
+                },
+                {
                   "type": "category",
                   "label": "Map",
                   "collapsed": true,
@@ -3114,6 +3151,10 @@ module.exports = {
                 },
                 {
                   "type": "doc",
+                  "id": "types/route"
+                },
+                {
+                  "type": "doc",
                   "id": "types/routeurlstrategy"
                 },
                 {
@@ -3608,6 +3649,10 @@ module.exports = {
                 },
                 {
                   "type": "doc",
+                  "id": "types/is_route_active"
+                },
+                {
+                  "type": "doc",
                   "id": "types/memo"
                 },
                 {
@@ -3641,6 +3686,26 @@ module.exports = {
                 {
                   "type": "doc",
                   "id": "types/useref"
+                },
+                {
+                  "type": "doc",
+                  "id": "types/use_route_loader_data"
+                },
+                {
+                  "type": "doc",
+                  "id": "types/use_route_location"
+                },
+                {
+                  "type": "doc",
+                  "id": "types/use_route_outlet"
+                },
+                {
+                  "type": "doc",
+                  "id": "types/use_route_params"
+                },
+                {
+                  "type": "doc",
+                  "id": "types/use_view_path"
                 },
                 {
                   "type": "doc",